requirements.txt
# Aurora Utensils Manufacturing – Project Dependencies

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
pyyaml>=6.0
openpyxl>=3.1.0
xlrd>=2.0.0
requests>=2.31.0
jupyter>=1.0.0
nbformat>=5.9.0
joblib>=1.3.0
